refactor: change timestamp type from int64_t to double in Budget and TruthValue classes

+++ nal/Budget.decl.hpp
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <cstdint>
-
 namespace seqnars::nal
 {
 struct Budget
@@ -14,10 +12,10 @@ struct Budget
     double priority;
     double durability;
     double quality;
-    int64_t ts_update;
+    double ts_update;
 
     Budget(double priority = PRIORITY_DEFAULT, double durability = DURABILITY_DEFAULT, double quality = QUALITY_DEFAULT,
-           int64_t ts_update = -1);
+           double ts_update = -1.0);
 
     static auto calc_durability(double half_life_period) -> double;
